+++ applications/temporalgrid-demo/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Patch Changes
 
+- Updated dependencies [ed81ea50]
+  - @globalfishingwatch/timebar@2.0.1
 - Updated dependencies [12482b76]
   - @globalfishingwatch/layer-composer@8.8.0